fix: enforce max-lines lint
