+++ packages/web/lib/github-webhook-handler.ts
@@ -1,4 +1,3 @@
-import { createHmac, timingSafeEqual } from "node:crypto";
 import type { IncomingMessage, ServerResponse } from "node:http";
 import type Database from "better-sqlite3";
 import {
@@ -9,6 +8,19 @@ import {
   recordWebhookEvent,
 } from "@issuectl/core";
 import type { WebhookTargetType } from "@issuectl/core";
+import {
+  asObject,
+  getBoundedHeader,
+  getNumberProperty,
+  getRepositoryFullName,
+  getSenderLogin,
+  getSingleHeader,
+  getStringProperty,
+  parseJson,
+  readRawBody,
+  verifySignature,
+  writeJson,
+} from "./github-webhook-utils.js";
 
 export const GITHUB_WEBHOOK_PATH_RE = /^\/api\/webhook\/github\/(\d+)$/;
 export const MAX_WEBHOOK_BODY_BYTES = 1024 * 1024;
@@ -27,8 +39,6 @@ const GATING_RELEVANT_EVENTS = new Set([
   "pull_request:synchronize",
   "pull_request:closed",
 ]);
-
-type JsonObject = Record<string, unknown>;
 
 export function isGithubWebhookRequest(url: string | undefined): boolean {
   return GITHUB_WEBHOOK_PATH_RE.test(
@@ -246,108 +256,6 @@ export function classifyWebhookTarget(payload: unknown): {
   }
 
   return nullTarget();
-}
-
-async function readRawBody(
-  req: IncomingMessage,
-  limitBytes: number,
-): Promise<{ buffer: Buffer; tooLarge: boolean }> {
-  const chunks: Buffer[] = [];
-  let total = 0;
-
-  for await (const chunk of req) {
-    const buffer = Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk);
-    total += buffer.length;
-    if (total > limitBytes) {
-      return { buffer: Buffer.alloc(0), tooLarge: true };
-    }
-    chunks.push(buffer);
-  }
-
-  return { buffer: Buffer.concat(chunks), tooLarge: false };
-}
-
-function verifySignature(
-  body: Buffer,
-  secret: string,
-  signature: string,
-): boolean {
-  const prefix = "sha256=";
-  if (!signature.startsWith(prefix)) return false;
-
-  const expected = createHmac("sha256", secret).update(body).digest();
-  const actualHex = signature.slice(prefix.length);
-  if (!/^[a-fA-F0-9]{64}$/.test(actualHex)) return false;
-
-  const actual = Buffer.from(actualHex, "hex");
-  return actual.length === expected.length && timingSafeEqual(actual, expected);
-}
-
-function writeJson(
-  res: ServerResponse,
-  statusCode: number,
-  body: Record<string, unknown>,
-): void {
-  res.statusCode = statusCode;
-  res.setHeader("content-type", "application/json; charset=utf-8");
-  res.end(JSON.stringify(body));
-}
-
-function getSingleHeader(
-  req: IncomingMessage,
-  name: string,
-): string | undefined {
-  const value = req.headers[name.toLowerCase()];
-  if (Array.isArray(value)) return value[0];
-  return value;
-}
-
-function getBoundedHeader(
-  req: IncomingMessage,
-  name: string,
-  maxLength: number,
-): string | null {
-  const value = getSingleHeader(req, name);
-  if (!value) return null;
-  const trimmed = value.trim();
-  if (trimmed.length === 0 || trimmed.length > maxLength) return null;
-  return trimmed;
-}
-
-function parseJson(body: Buffer): JsonObject | null {
-  try {
-    return asObject(JSON.parse(body.toString("utf8")));
-  } catch {
-    return null;
-  }
-}
-
-function getRepositoryFullName(payload: JsonObject): string | null {
-  return getStringProperty(asObject(payload.repository), "full_name");
-}
-
-function getSenderLogin(payload: JsonObject): string | null {
-  return getStringProperty(asObject(payload.sender), "login");
-}
-
-function asObject(value: unknown): JsonObject | null {
-  if (!value || typeof value !== "object" || Array.isArray(value)) return null;
-  return value as JsonObject;
-}
-
-function getStringProperty(
-  object: JsonObject | null,
-  key: string,
-): string | null {
-  const value = object?.[key];
-  return typeof value === "string" ? value : null;
-}
-
-function getNumberProperty(object: JsonObject, key: string): number | null {
-  const value = object[key];
-  return typeof value === "number" && Number.isInteger(value) && value > 0
-    ? value
-    : null;
 }
 
 function nullTarget(): {
